Development (#64)
* Using mock sponsors that have pictures

* Use env variables if present

Use env variables if present otherwise just assume the defaults.

* Update to use random data

Use uuid as the name of sponsor, players etc

* Remove the scripts

They are not really needed and brought up the question of supporting windows (bat files). Additionally, they were not tested on linux so guessing that winpty only really worked for git bash.

* Just compare the date strings

If it before or game is today then give the score as 0-0

* Remove some un-used routes

The one line was being overwritten so it is not needed

* Unit test for test view

Additionally, found a few issue with some of the routes. Most of it was just safe guarding against cases where do not find an object

* 404 code constant

* Fixed up the spacing

Additionally, remove some un-used maps

* Fixed the function

Discovered the issue when writing a unit-test

* Add some more tests

Got to 85% for views which is good enough for now

* Remove the captain stuff

It is not tested and not used

* Add a simple test for documentation

* Add in coverage reporting

fail if fails under 80%

* Product a coverage report

* The action already zips it

No need to zip

* Update gitignore for cypress project

Have it ignore certain node things

* Add to data-cy attributes for testing

* Redirect to homepage no index

Since index requires a year

* Try calling ready and catch any errors

Checking if it was defined did not work when it was

* Add some data-cy selectors

* First commit for cypress testing

Leaving selenium for now but will eventually move all the tests over.

* A simple cypress feature test for login

* Try run in-memory server and running cypress

* Fix action to run properly

Use combined feature files & pass admin/password as env

* Fix gitignore and add combines features files for admin & website (#67)

* Update the action to try and run cypress testing

Additionally, upload the artifact

* Ignore vdeios from admin and website

* Missing starting path for ignore videos and screenshots

* Combine all features into one file

This allows for one recording to be uploaded to website and for admin

* Just run the combined feature files

Pass in the variables use --env as well