Make the model a fair bit easier to work with
The colum property allows for the model to handle most of the logic with summing game scores and team standings.
Further work can be done here. This makes it much easier to just use the model and allow it to handle the optimal SQL queries.
Added a simple test and update team_stats to use the properties. Next step is to use it more player stats and for team stats. This should
speed up a bunch of things.