Add division table (#72)
* Added error and error handler for divsion does not exist

* Fixed a typo in the documentation

* Update model.py

* Added a test for division model

Also made league model test use random strings

* First attempt at a division category

* Add the new division API to a route

* Fix parameter name and some doc

* Fix a typo

* Added the division route

* The testing of the division API

* Update the game model to expose the division id

Fix the test and documentation

* Update the game model to use division id

* Make Division required for games

Fix most of the unit tests except for the import
I think the import template should be updated to not include the headers but a drop down
before the header and part of the request.

* Remove the debugging

* Set the time such that near the last hour of the year.

* Mock the league so that tests will run at end and beginning of a year

* Fix the assertion message

* Require division when importing a list of games

* Update the games template to include division

* Fix the tests to now work with division

* Update test tempaltes to use 2020 and division

* Refactor the code to make it simpler to read

* Add division id to viewing the results of games

This allows to sort the game scores by division id

* Update documentation to include division and being part of game

* Add documentation routes for division

* Division is required by game now

* Model was not using the given division id

* Make the schedule page more responsive for mobile users

I imagine this is the most used from a mobile phone

* Make the stats page show SS and HR over the over stats

Help make the website a little more responsive

* Use the responsive attributes

* Set the responsive attributes in the header

* Fix a couple of small bugs where changing onclick instead of onchange

* onchange instead of onclick

* Get the page for editing division working

* Seems cypress has issues with Ubuntu Latest

https://github.com/cypress-io/github-action

* Include the division in the response

* The put was not updating the division id

* Do not want team id to in the division table

* Forgetting to drop some tables

* Getting clode to letting edit game to work

* Handle the case where either team id is None

This was happening from the front ent so seems possible

* Clear the data upon success

Validate the data a little more on the front end before sending to backend

* Make the model a fair bit easier to work with

The colum property allows for the model to handle most of the logic with summing game scores and team standings.
Further work can be done here. This makes it much easier to just use the model and allow it to handle the optimal SQL queries.
Added a simple test and update team_stats to use the properties. Next step is to use it more player stats and for team stats. This should
speed up a bunch of things.

* This API now returns espys

* Update the website to handle the changes for division

Still need for a way to sort by division for the schedule and for the standings as well.

* Remove espys_awarded.

* Fix the unit tests

* Fix cypress test

* fix the request url